LPS-53183 Move code to common place and use osgi API's
