+++ modules/core/servicetrackermap/src/com/liferay/osgi/servicetrackermap/ServiceReferenceServiceTuple.java
@@ -12,9 +12,9 @@
  * details.
  */
 
-package com.liferay.registry.collections;
+package com.liferay.osgi.servicetrackermap;
 
-import com.liferay.registry.ServiceReference;
+import org.osgi.framework.ServiceReference;
 
 /**
  * @author Carlos Sierra Andrés
