+++ modules/core/servicetrackermap/src/com/liferay/osgi/servicetrackermap/internal/SingleValueServiceTrackerBucketFactory.java
@@ -12,17 +12,18 @@
  * details.
  */
 
-package com.liferay.registry.collections.internal;
+package com.liferay.osgi.servicetrackermap.internal;
 
-import com.liferay.registry.ServiceReference;
-import com.liferay.registry.collections.ServiceReferenceServiceTuple;
-import com.liferay.registry.collections.ServiceReferenceServiceTupleComparator;
-import com.liferay.registry.collections.ServiceTrackerBucket;
-import com.liferay.registry.collections.ServiceTrackerBucketFactory;
+import com.liferay.osgi.servicetrackermap.ServiceReferenceServiceTuple;
+import com.liferay.osgi.servicetrackermap.ServiceReferenceServiceTupleComparator;
+import com.liferay.osgi.servicetrackermap.ServiceTrackerBucket;
+import com.liferay.osgi.servicetrackermap.ServiceTrackerBucketFactory;
 
 import java.util.Collections;
 import java.util.Comparator;
 import java.util.PriorityQueue;
+
+import org.osgi.framework.ServiceReference;
 
 /**
  * @author Carlos Sierra Andrés
